refactor(config): lua/abdellatifdev -> lua/adev
