--- lua/abdellatifdev/plugins/comments.lua
@@ -1,5 +1,0 @@
-return {
-    'numToStr/Comment.nvim',
-    event = require("abdellatifdev.consts").events.pre,
-    opts = {}
-}
