+++ lua/adev/plugins/blink-cmp.lua
@@ -7,7 +7,7 @@ return {
     },
     {
         'saghen/blink.cmp',
-        event = require("abdellatifdev.consts").events.insert,
+        event = require("adev.consts").events.insert,
         version = "v1.*",
         dependencies = {
             'saghen/blink.compat',
