+++ lua/adev/plugins/comments.lua
@@ -1,0 +1,5 @@
+return {
+    'numToStr/Comment.nvim',
+    event = require("adev.consts").events.pre,
+    opts = {}
+}
