+++ lua/adev/plugins/git-conflict.lua
@@ -5,5 +5,5 @@ return {
         return vim.fn.system("git rev-parse --is-inside-work-tree 2>/dev/null"):gsub("\n", "") == "true"
     end,
     opts = {},
-    event = require("abdellatifdev.consts").events.file,
+    event = require("adev.consts").events.file,
 }
