+++ lua/adev/plugins/none-ls.lua
@@ -1,0 +1,5 @@
+return {
+    'nvimtools/none-ls.nvim',
+    event = require("adev.consts").events.file,
+    config = function() require('adev.config.none-ls').setup() end
+}
