+++ lua/adev/plugins/treesitter.lua
@@ -20,7 +20,7 @@ end
 return {
     {
         "nvim-treesitter/nvim-treesitter-context",
-        event = require("abdellatifdev.consts").events.file,
+        event = require("adev.consts").events.file,
         opts = {
             enable = true,
             max_lines = 2,
@@ -30,7 +30,7 @@ return {
     {
         'windwp/nvim-ts-autotag',
 
-        event = require("abdellatifdev.consts").events.file,
+        event = require("adev.consts").events.file,
         opts = {
             per_filetype = {
                 ["blade"] = {
@@ -48,7 +48,7 @@ return {
     },
     {
         "nvim-treesitter/nvim-treesitter",
-        event = require("abdellatifdev.consts").events.file,
+        event = require("adev.consts").events.file,
         build = ":TSUpdate",
         config = treesitter_setup,
     },
